+++ changelog.md
@@ -53,3 +53,7 @@
 * Fix incorrect `Equals` implementation on `Option`.
   `Equals` previously returned `true` when comparing a `None` value with a `Some` value containing the default value of the type.
 * Add `None` extension method to `IEnumerable`.
+* `Option<Task<T>>`, `Option<Task>` and their `ValueTask` equivalents are now awaitable:
+  ```csharp
+  var answer = await Option.Some(Task.FromResult(42));
+  ```